requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2.20
frozendict==2.4.6
idna==3.10
multitasking==0.0.11
numpy==2.0.2
pandas==2.2.3
peewee==3.17.9
platformdirs==4.3.6
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
yfinance==0.2.54
django-cors-headers
djangorestframework
